refactor: imageupload component
